+++ RELEASES.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+## [Unreleased]
+
+### Fixed
+
+* rapl [#46]: Check that top-level zones are actually package (not psys) zones
+
 ## [v0.3.0] - 2020-03-10
 
 ### Added
@@ -116,11 +122,13 @@
 * Initial public release
 
 
+[Unreleased]: https://github.com/energymon/energymon/compare/v0.3.0...HEAD
 [v0.3.0]: https://github.com/energymon/energymon/compare/v0.2.4...v0.3.0
 [v0.2.4]: https://github.com/energymon/energymon/compare/v0.2.3...v0.2.4
 [v0.2.3]: https://github.com/energymon/energymon/compare/v0.2.2...v0.2.3
 [v0.2.2]: https://github.com/energymon/energymon/compare/v0.2.1...v0.2.2
 [v0.2.1]: https://github.com/energymon/energymon/compare/v0.2.0...v0.2.1
+[#46]: https://github.com/energymon/energymon/issues/46
 [#44]: https://github.com/energymon/energymon/issues/44
 [#40]: https://github.com/energymon/energymon/issues/40
 [#37]: https://github.com/energymon/energymon/issues/37